test modules, do not use it

+++ main.py
@@ -46,7 +46,7 @@ def main(argv):
         elif(opt in ("-p","--prob")):
             splittedInput = arg.split(",")
             for elem in splittedInput:
-                p.append(int(elem))
+                p.append(float(elem))
         elif (opt in ("-r", "--rate")):
             splittedInput = arg.split(",")
             for elem in splittedInput:
